backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
python-dotenv==1.0.0

# AI/ML Core
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.1
transformers==4.36.0
torch>=2.0.0
sentence-transformers==2.2.2
langchain==0.0.350
langchain-community==0.0.1
langchain-openai==0.0.2
langchain-google-genai==0.0.6

# Vector Databases & Search
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18
faiss-cpu==1.7.4

# Web Search & Scraping
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
duckduckgo-search==3.9.6
arxiv==1.4.8
scholarly==1.7.11
serpapi==0.1.5
googlesearch-python==1.2.3

# Memory & Caching
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.0

# Multimodal
Pillow==10.1.0
opencv-python==4.8.1.78
openai-whisper==20231117
moviepy==1.0.3
pytesseract==0.3.10
easyocr==1.7.0
speechrecognition==3.10.0
pydub==0.25.1
librosa==0.10.1

# Security & Privacy
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
numpy==1.25.2
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-dateutil==2.8.2
tiktoken==0.5.2
tenacity==8.2.3
asyncio-throttle==1.0.2
websockets==12.0
sse-starlette==1.6.5

# Advanced Features
spacy==3.7.2
nltk==3.8.1
textstat==0.7.3
yake==0.4.8
keybert==0.8.3
sumy==0.11.0
newspaper3k==0.2.8